Create separate html submodule in wasm.
