+++ ftml/src/wasm/render/html.rs
@@ -1,5 +1,5 @@
 /*
- * wasm/render.rs
+ * wasm/render/html.rs
  *
  * ftml - Library to parse Wikidot text
  * Copyright (C) 2019-2022 Wikijump Team
@@ -18,22 +18,21 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-use super::page_info::PageInfo;
-use super::parsing::SyntaxTree;
-use super::prelude::*;
-use super::settings::WikitextSettings;
-use crate::render::text::TextRender;
-use crate::render::Render;
+//! Isolated module for WASM HTML rendering.
+//!
+//! This submodule is separate to easily gate it within `#[cfg(feature = "html")]`,
+//! and so imports essentially the same fields as its parent.
 
-#[cfg(feature = "html")]
-use std::sync::Arc;
-
-#[cfg(feature = "html")]
+use super::super::page_info::PageInfo;
+use super::super::parsing::SyntaxTree;
+use super::super::prelude::*;
+use super::super::settings::WikitextSettings;
 use crate::render::html::{HtmlOutput as RustHtmlOutput, HtmlRender};
+use crate::render::Render;
+use std::sync::Arc;
 
 // Typescript declarations
 
-#[cfg(feature = "html")]
 #[wasm_bindgen(typescript_custom_section)]
 const TS_APPEND_CONTENT: &str = r#"
 
@@ -57,7 +56,6 @@ export interface IBacklinks {
 
 "#;
 
-#[cfg(feature = "html")]
 #[wasm_bindgen]
 extern "C" {
     #[wasm_bindgen(typescript_type = "string[]")]
@@ -72,14 +70,12 @@ extern "C" {
 
 // Wrapper structures
 
-#[cfg(feature = "html")]
 #[wasm_bindgen]
 #[derive(Debug, Clone)]
 pub struct HtmlOutput {
     inner: Arc<RustHtmlOutput>,
 }
 
-#[cfg(feature = "html")]
 #[wasm_bindgen]
 impl HtmlOutput {
     #[wasm_bindgen]
@@ -110,9 +106,8 @@ impl HtmlOutput {
     }
 }
 
-// Exported functions
+// Function exports
 
-#[cfg(feature = "html")]
 #[wasm_bindgen]
 pub fn render_html(
     syntax_tree: SyntaxTree,
@@ -127,18 +122,4 @@ pub fn render_html(
     HtmlOutput {
         inner: Arc::new(html),
     }
-}
-
-#[wasm_bindgen]
-pub fn render_text(
-    syntax_tree: SyntaxTree,
-    page_info: PageInfo,
-    settings: WikitextSettings,
-) -> String {
-    let tree = syntax_tree.get();
-    let page_info = page_info.get();
-    let settings = settings.get();
-    let text = TextRender.render(tree, page_info, settings);
-
-    text
 }
